+++ src/components/GuestsPaymentModal/index.js
@@ -1,0 +1,2 @@
+import GuestsPaymentModal from './GuestsPaymentModal';
+export default GuestsPaymentModal;
